src/app/shared/select/select.component.ts
import { NgFor } from '@angular/common';
import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';

@Component({
  selector: 'app-select',
  standalone: true,
  imports: [NgFor],
  templateUrl: './select.component.html',
  styleUrl: './select.component.scss',
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class SelectComponent {
  @Input() data: number[] = [];
  @Output() selectionChange = new EventEmitter<number>();

  emitSelectChange(value: number) {
    this.selectionChange.emit(value);
  }
}
